fix data directory startup using old DB

+++ setup-database.sh
@@ -8,17 +8,19 @@ SETUP_LOCKFILE="${DATADIR}/.postgresql.init.lock"
 # This script will setup the necessary folder for database
 
 # test if DATADIR has content
-if [[ ! -f "${SETUP_LOCKFILE}" ]]; then
-  # No content yet - first time pg is being run!
-  # No Replicate From settings. Assume that this is a master database.
-  # Initialise db
-  echo "Initializing Postgres Database at ${DATADIR}"
-  rm -rf ${DATADIR}
-  #chown -R postgres $DATADIR
-  su - postgres -c "$INITDB -E ${DEFAULT_ENCODING} --lc-collate=${DEFAULT_COLLATION} --lc-ctype=${DEFAULT_CTYPE} --wal-segsize=${WAL_SEGSIZE} ${DATADIR}"
-  touch ${SETUP_LOCKFILE}
-fi
+if [[ -z "${EXISTING_DATA_DIR}" ]]; then \
+    if [[ ! -f "${SETUP_LOCKFILE}" ]]; then
+        # No content yet - first time pg is being run!
+        # No Replicate From settings. Assume that this is a master database.
+        # Initialise db
+        echo "Initializing Postgres Database at ${DATADIR}"
+        rm -rf ${DATADIR}
+        #chown -R postgres $DATADIR
+        su - postgres -c "$INITDB -E ${DEFAULT_ENCODING} --lc-collate=${DEFAULT_COLLATION} --lc-ctype=${DEFAULT_CTYPE} --wal-segsize=${WAL_SEGSIZE} ${DATADIR}"
+        touch ${SETUP_LOCKFILE}
+    fi
 
+fi;
 # Set proper permissions
 # needs to be done as root:
 chown -R postgres:postgres ${DATADIR}
